Save id dropdown for realm debugging.

+++ Yume/Assets/_Main/_Scripts/Realm/DebuggingRealm.cs
@@ -3,5 +3,5 @@ using Realms;
 public class DebuggingRealm : RealmObject
 {
     public string SceneName { get; set; }
-    public int SaveId { get; set; }
+    public long SaveId { get; set; }
 }
